requirements.txt
apache-airflow
kafka
kafka-python
mysql-connector-python